chore: rename SearchSections into SearchSection
