+++ frontend/src/components/Recipes/Sidebar/SearchSection/SearchSection.tsx
