+++ frontend/src/components/Recipes/Sidebar/SearchSection/SelectCategory.tsx
