+++ frontend/src/components/Recipes/Sidebar/SearchSection/SelectCuisine.tsx
