+++ frontend/src/components/Recipes/Sidebar/index.ts
@@ -2,7 +2,7 @@ import RecipeFilters from './FiltersSection/RecipeFilters'
 import IngredientsList from './FiltersSection/IngredientsList'
 import UserReviews from './FiltersSection/UserReviews'
 import RecipeDifficulty from './FiltersSection/RecipeDifficulty'
-import SearchSections from './SearchSections/SearchSections'
+import SearchSections from './SearchSection/SearchSection'
 
 export {
     SearchSections,
